Creating service for status
