+++ packages/core/base/src/index.ts
@@ -2,3 +2,4 @@ export * from "./utils/index";
 export * from "./models/types";
 export * from "./services/crossmintModalService";
 export * from "./services/crossmintPayButtonService";
+export * from "./services/crossmintStatusService";
